requirements.txt
PyQt6==6.6.1
Pillow==10.1.0
cx-Freeze==6.15.12
pytest>=7.0.0   # For testing
black>=23.0.0   # For code formatting
flake8>=6.0.0   # For linting